+++ paddle/fluid/inference/api/paddle_pass_builder.cc
@@ -526,7 +526,8 @@ XpuPassStrategy::XpuPassStrategy() : PassStrategy({}) {
       "one_beam_size_fuse_pass",
       "fold_interp_outsize_fuse_pass",
       "fold_two_squeeze2_fuse_pass",
-      "reduce_max_fuse_pass",
+      "redundant_onnx_ops_elimination_pass",
+      "reduce_ops_fuse_pass",
       "delete_cast_op_pass",
       "xpu_delete_cast_op_pass",
       "stack_fuse_pass",
@@ -538,7 +539,6 @@ XpuPassStrategy::XpuPassStrategy() : PassStrategy({}) {
       "map_matmulv2_to_matmul_xpu_pass",
       "reshape2_matmul_xpu_fuse_pass",
       "squeeze2_matmul_xpu_fuse_pass",
-      "redundant_onnx_ops_elimination_pass",
       "redundant_squeeze_unsqueeze_elimination_pass",
       "fc_xpu_fuse_pass",
       "conv2d_xpu_fuse_pass",
